Have Tiny Server run on a separate thread
The original motivation for this change was that I wanted to
setup a separate struct for the Hello World Server. Something like this:

```
pub struct HelloServer {
	tiny_server: TinyServer,
	foo: u32,
	bar: u32,
}

impl HelloServer {
	pub fn main(self) -> ! {
		loop {
			self.foo += 1;
			self.tiny_server.poll();
			self.bar += 1;
		}
	}
}
```

This did not compile. The issue is with the callback that's passed into
`TinyServer`'s constructor. As a result, for `foo: HelloServer`,
`foo.tiny_server` contains a reference to `foo` itself. This caused all sorts
of issus with the borrow checker, and rightly so. I experimented with the idea
of having the callback passed as a paramter to `TinyServer`'s `poll` function
(instead of `TinyServer`'s constructor), but that also caused compiler errors
because `self` was mutably borrowed once to construct `poll`'s arguments, but
it then needed to be borrowed mutably again when actually calling
`self.tiny_server.poll`.

I tried a few other things as well, before deciding to go with the solution
from this commit: entirely get rid of the `TinyServer` struct, and have that
logic run on a separate thread instead.